pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "nltk>=3.9.1",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "spacy>=3.8.4",
    "streamlit>=1.41.1",
]
